+++ jarvis/executors/unconditional.py
@@ -11,7 +11,7 @@ from jarvis.modules.conditions import keywords
 from jarvis.modules.exceptions import EgressErrors
 from jarvis.modules.logger.custom_logger import logger
 from jarvis.modules.models import models
-from jarvis.modules.utils import util
+from jarvis.modules.utils import support
 
 
 def google_maps(query: str) -> bool:
@@ -92,8 +92,8 @@ def google_maps(query: str) -> bool:
         speaker.speak(text=f"The {option}, {item['Name']}, with {item['Rating']} rating, "
                            f"on{''.join([j for j in item['Address'] if not j.isdigit()])}, which is approximately "
                            f"{dist} away. {next_val}", run=True)
-        util.write_screen(text=f"{item['Name']} -- {item['Rating']} -- "
-                               f"{''.join([j for j in item['Address'] if not j.isdigit()])}")
+        support.write_screen(text=f"{item['Name']} -- {item['Rating']} -- "
+                                  f"{''.join([j for j in item['Address'] if not j.isdigit()])}")
         if converted := listener.listen():
             if 'exit' in converted or 'quit' in converted or 'Xzibit' in converted:
                 break